json_gen_str_end: Return total length of the JSON created
Even if json_gen_str_start() is now called with a NULL buffer,
the subsequent APIs keep a track of the total length (even though
the JSON itself won't be created), so that at the end, when
json_gen_str_end() is called, the total length required for the
JSON can be found. Applications can then allocate a buffer equal
to this size and then call all the APIs again.